# backend-java/docs/tenant-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI-Bid 多租户后端 API",
    "version": "tenant-v1",
    "description": "Java 业务网关的多租户认证、租户、成员、邀请和审核 SSE 契约。租户 ID 在 JSON、JWT、消息和 Header 中使用十进制字符串。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/tenants": {
      "get": {
        "operationId": "listTenants",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TenantPage"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createTenant",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Tenant"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/tenants/current": {
      "get": {
        "operationId": "getCurrentTenant",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TenantSummary"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/tenants/{tenantId}/members": {
      "get": {
        "operationId": "listMembers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemberPage"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "TenantId": {
        "name": "tenantId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[0-9]+$"
        }
      },
...
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Size": {
        "name": "size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "schemas": {
      "Result": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "msg": {
            "type": "string"
          },
          "data": {},
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ErrorData": {
        "type": "object",
        "required": [
          "error_code",
          "request_id"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ErrorData"
              }
            }
          }
        ]
      },
...
      "TenantRole": {
        "type": "string",
        "enum": [
          "OWNER",
          "ADMIN",
          "AUDITOR",
          "MEMBER",
          "VIEWER"
        ]
      },
      "TenantSummary": {
        "type": "object",
        "required": [
          "tenant_id",
          "tenant_code",
          "name",
          "status",
          "role",
          "permissions",
          "is_current"
        ],
        "properties": {
          "tenant_id": {
            "type": "string"
          },
          "tenant_code": {
            "type": "string",
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "maxLength": 128
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "DELETED"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "is_current": {
            "type": "boolean"
          }
        }
      },
      "Tenant": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TenantSummary"
          },
          {
            "type": "object",
            "required": [
              "owner_user_id",
              "plan_code",
              "version",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "owner_user_id": {
                "type": "string"
              },
              "plan_code": {
                "type": "string"
              },
              "settings": {
                "type": "object",
                "additionalProperties": true,
                "nullable": true
              },
              "version": {
                "type": "integer",
                "format": "int64"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
...
      "TenantMember": {
        "type": "object",
        "required": [
          "member_id",
          "tenant_id",
          "user_id",
          "role",
          "permissions",
          "status",
          "joined_at"
        ],
        "properties": {
          "member_id": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "real_name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "REMOVED"
            ]
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "CreateTenantRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "tenant_code": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{2,63}$"
          },
          "plan_code": {
            "type": "string",
            "default": "STANDARD"
          },
          "settings": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    },
    "responses": {
...
      "Tenant": {
        "description": "租户对象",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Result"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Tenant"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "TenantSummary": {
        "description": "当前租户摘要",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Result"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TenantSummary"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "TenantPage": {
        "description": "租户列表",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Result"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "current_tenant_id",
                        "items"
                      ],
                      "properties": {
                        "current_tenant_id": {
                          "type": "string",
                          "nullable": true
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TenantSummary"
                          }
                        }
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "MemberPage": {
        "description": "成员列表",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Result"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "size": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TenantMember"
                          }
                        }
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "Error": {
        "description": "统一错误响应",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
...
}